Move PEC enforcement to enterprise
