+++ enterprise/ha/src/main/java/org/neo4j/kernel/ha/factory/PropertyExistenceEnforcer.java
@@ -5,20 +5,19 @@
  * This file is part of Neo4j.
  *
  * Neo4j is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * it under the terms of the GNU Affero General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
+ * GNU Affero General Public License for more details.
  *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-
-package org.neo4j.kernel.impl.api;
+package org.neo4j.kernel.ha.factory;
 
 import java.util.Iterator;
 import java.util.Set;
@@ -45,7 +44,7 @@ import org.neo4j.kernel.impl.api.store.StoreStatement;
 
 import static java.lang.String.format;
 
-public class PropertyExistenceEnforcer extends TxStateVisitor.Adapter
+class PropertyExistenceEnforcer extends TxStateVisitor.Adapter
 {
     private final EntityReadOperations readOperations;
     private final StoreReadLayer storeLayer;
@@ -82,7 +81,7 @@ public class PropertyExistenceEnforcer extends TxStateVisitor.Adapter
         validateNode( id );
         super.visitNodeLabelChanges( id, added, removed );
     }
-    
+
     @Override
     public void visitCreatedRelationship( long id, int type, long startNode, long endNode )
             throws ConstraintValidationKernelException
@@ -98,8 +97,6 @@ public class PropertyExistenceEnforcer extends TxStateVisitor.Adapter
         validateRelationship( id );
         super.visitRelPropertyChanges( id, added, changed, removed );
     }
-    
-
 
     private void validateNode( long nodeId ) throws ConstraintValidationKernelException
     {
